requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
python-dotenv==1.0.1
